requirements.txt
pyrogram>=2.0.106
tgcrypto==1.2.5
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
alembic==1.12.1

